requirements.txt
MetaTrader5
pyTelegramBotAPI
python-dotenv
pandas
numpy
schedule
pytest
black
